migrate changes to new files / folders
